Turn this crate into a library
