+++ Cargo.toml
@@ -9,5 +9,7 @@ bindgen = "~0.42.1"
 
 [dependencies]
 libc = "0.2"
-net-literals = "0.1.2"
 nix = "0.11"
+
+[dev-dependencies]
+net-literals = "0.1.2"
