+++ src/lib.rs
@@ -1,0 +1,8 @@
+//! This crate provides libutp Rust bindings.
+
+extern crate libc;
+extern crate nix;
+
+mod utp;
+
+pub use utp::*;
